Split templateProcessor with a common part and a specific to Word
